feat: Add shiftByUser route and view
The code changes in this commit add a new route and view for displaying shifts by user. The `shiftByUser` function is added to the `shiftController.js` file to handle the logic of retrieving shifts by user ID. The corresponding route is updated in the `shift.js` file to use the `shiftByUser` function. The `shiftByUser.pug` view is also created to render the shifts in a table format. This feature enhances the user experience by allowing users to view their shifts.